Milestone/2.16

+++ app/src/main/java/com/ruuvi/station/database/tables/RuuviTagEntity.kt
@@ -69,7 +69,7 @@ data class RuuviTagEntity(
     @Column
     var nox: Int? = null,
     @Column
-    var luminosity: Int? = null,
+    var luminosity: Double? = null,
     @Column
     var dBaAvg: Double? = null,
     @Column
